+++ examples/resources/graylog_output/resource.tf
@@ -1,6 +1,6 @@
 resource "graylog_output" "example" {
   title = "Terraform Output"
-  type  = "org.graylog2.outputs.LoggingOutput"
+  type  = "LoggingOutput"
   configuration_json = jsonencode({
     prefix = "terraform-output:"
   })